Remove unused api models from common module

+++ accounts/src/main/java/org/vl4ds4m/banking/accounts/messaging/AccountBrokerMessage.java
@@ -1,11 +1,11 @@
 package org.vl4ds4m.banking.accounts.messaging;
 
-import org.vl4ds4m.banking.currency.Currency;
+import org.vl4ds4m.banking.accounts.entity.Currency;
 
 import java.math.BigDecimal;
 
 public record AccountBrokerMessage(
-    Integer accountNumber,
+    Long accountNumber,
     Currency currency,
     BigDecimal balance
 ) {
